+++ R/fviz_ca.R
@@ -8,7 +8,7 @@ NULL
 #'  visualization of CA outputs from the R functions: CA [in FactoMineR], ca [in
 #'  ca], coa [in ade4], correspondence [in MASS] and expOutput/epCA [in
 #'  ExPosition]. Read more: 
-#'  \href{http://www.sthda.com/english/wiki/correspondence-analysis-in-r-the-ultimate-guide-for-the-analysis-the-visualization-and-the-interpretation-r-software-and-data-mining}{Correspondence
+#'  \href{https://www.sthda.com/english/articles/31-principal-component-methods-in-r-practical-guide/113-ca-correspondence-analysis-in-r-essentials/}{Correspondence
 #'   Analysis}
 #'  
 #'  \itemize{ \item{fviz_ca_row(): Graph of row variables} \item{fviz_ca_col():