+++ R/fviz_silhouette.R
@@ -11,7 +11,7 @@
 #'  factoextra].
 #'  
 #'  Read more: 
-#'  \href{http://www.sthda.com/english/wiki/clustering-validation-statistics-4-vital-things-everyone-should-know-unsupervised-machine-learning}{Clustering
+#'  \href{https://www.datanovia.com/en/lessons/cluster-validation-statistics-must-know-methods/}{Clustering
 #'   Validation Statistics}.
 #'@details - Observations with a large silhouhette Si (almost 1) are very well 
 #'  clustered.