+++ NEWS.md
@@ -1,4 +1,4 @@
-# factoextraplus (unreleased)
+# factoextraplus 1.2.0
 
 ## Breaking changes